Separate parser bounds checks from mode checks.
Whether collectingContent is -1 determines if the parser mode: if it is
collecting headers or body data.

The secondary check in each if-statement looked for the data boundary.
This commit separates these checks, because they will become our loop
terminators.

Additionally, this prevents buffer indexing for the header terminator
when we're not parsing headers.

This commit in itself fails tests, because headers and body often arrive
in a single write, and the parser stalls.